Remove package installation tests, to be reintroduced at citusdata/packaging#1171
